Improved onStart setting test
