--- test/programs/onstart-bad/cli/entry.js
@@ -1,9 +1,0 @@
-// Import and initialize Waterfall CLI
-const waterfall = require('../../../../dist/index');
-
-const onStart = 'This is not a function';
-
-waterfall.init({
-	verbose: true,
-	onStart: onStart,
-});
