--- test/programs/onstart-bad/cli/entry.json
@@ -1,1 +1,0 @@
-{}
