--- test/programs/onstart-bad/package.json
@@ -1,4 +1,0 @@
-{
-	"name": "onstart-bad",
-	"version": "1.2.3"
-}
