--- test/programs/onstart-good/cli/entry.json
@@ -1,1 +1,0 @@
-{}
