--- test/programs/onstart-good/cli/list/list.js
@@ -1,8 +1,0 @@
-// Require Waterfall CLI
-const waterfall = require('../../../../../dist/index');
-
-// Parse command input
-const input = waterfall.command();
-
-// Output result
-console.log(input);
