--- test/programs/onstart-good/package.json
@@ -1,4 +1,0 @@
-{
-	"name": "onstart-good",
-	"version": "1.2.3"
-}
