+++ test/programs/onstart-setting/cli/entry.json
@@ -1,0 +1,19 @@
+{
+	"flags": {
+		"quiet": {
+			"cascades": true,
+			"shorthand": "q",
+			"description": "Disable interactivity, rely on default values instead"
+		},
+		"non-cascading": {
+			"description": "Just used for testing"
+		}
+	},
+	"options": {
+		"delivery-zip-code": {
+			"cascades": true,
+			"shorthand": "z",
+			"description": "The delivery ZIP code, for context"
+		}
+	}
+}
