+++ test/programs/onstart-setting/cli/list/list.js
