+++ test/programs/onstart-setting/package.json
@@ -1,0 +1,4 @@
+{
+	"name": "onstart-setting",
+	"version": "1.2.3"
+}
